Prune ADR027 of post-decision speculation (#488)
This PR removes some speculative reasons from ADR027. These speculative reasons were added since the reasons why the selected option was chosen in ADR027 are unclear and not recorded in project documentation. Since the speculation would not accurately reflect the aspects considered when ADR027 was decided, Aaron and I decided it should be removed.